+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Adding `c::varToCamelCase()`.
 - Adding `c::activeSelected()`.
+- Adding CSS filter shadow utilities.
+- Optimizing CSS utilities that set unit values.
 - Bug fix. `curl_errno` should not impact the `http_code` return value.
 
 ## v170329.13807